CustomBottomNavigation guardando estado terminado
