+++ lib/config/router/app_router.dart
@@ -2,22 +2,31 @@ import 'package:cinemapedia/presentation/screens/screens.dart';
 import 'package:go_router/go_router.dart';
 
 final appRouter = GoRouter(
-  initialLocation: '/',
+  initialLocation: '/home/0',
   routes: [
     GoRoute(
-        path: '/',
-        name: HomeScreen.name,
-        builder: (context, state) => const HomeScreen(),
-        routes: [
-          GoRoute(
-            path: 'movie/:id',
-            name: MovieScreen.name,
-            builder: (context, state) {
-              final movieId = state.pathParameters['id'] ?? 'no-id';
+      path: '/home/:page',
+      name: HomeScreen.name,
+      builder: (context, state) {
+        final pageIndex = int.parse(state.pathParameters['page'] ?? '0');
 
-              return MovieScreen(movieId: movieId);
-            },
-          ),
-        ]),
+        return HomeScreen(pageIndex: pageIndex);
+      },
+      routes: [
+        GoRoute(
+          path: 'movie/:id',
+          name: MovieScreen.name,
+          builder: (context, state) {
+            final movieId = state.pathParameters['id'] ?? 'no-id';
+
+            return MovieScreen(movieId: movieId);
+          },
+        ),
+      ],
+    ),
+    GoRoute(
+      path: '/',
+      redirect: (_, __) => '/home/0',
+    ),
   ],
 );
