+++ lib/presentation/widgets/movies/movie_horizontal_listview.dart
@@ -108,7 +108,7 @@ class _Slide extends StatelessWidget {
                   }
 
                   return GestureDetector(
-                    onTap: () => context.push('/movie/${movie.id}'),
+                    onTap: () => context.push('/home/0/movie/${movie.id}'),
                     child: FadeInRight(child: child),
                   );
                 },
