+++ lib/presentation/widgets/movies/movies_slideshow.dart
@@ -70,7 +70,7 @@ class _Slide extends StatelessWidget {
               }
 
               return GestureDetector(
-                onTap: () => context.push('/movie/${movie.id}'),
+                onTap: () => context.push('/home/0/movie/${movie.id}'),
                 child: FadeIn(child: child),
               );
             },
